+++ ado_wrapper/state_managed_abc.py
@@ -3,7 +3,9 @@ from dataclasses import dataclass, fields
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, Callable, Literal, Type, TypeVar
 
-from ado_wrapper.errors import DeletionFailed, ResourceAlreadyExists, ResourceNotFound, UnknownError, UpdateFailed, InvalidPermissionsError  # fmt: skip
+from ado_wrapper.errors import (
+    DeletionFailed, ResourceAlreadyExists, ResourceNotFound, UnknownError, UpdateFailed, InvalidPermissionsError
+)  # fmt: skip
 from ado_wrapper.utils import extract_id, get_internal_field_names, get_resource_variables  # , get_editable_fields
 
 if TYPE_CHECKING:
@@ -121,45 +123,6 @@ class StateManagedResource:
         ado_client.state_manager.add_resource_to_state(resource)
         return resource  # [return-value]
 
-    # @classmethod
-    # def maintain(cls: Type[T], ado_client: "AdoClient", *args, **kwargs) -> T:  # url: str, payload: dict[str, Any]
-    #     # To be able to get the existing one, we need to be able to fetch it by name
-    #     # We can assume that all the resources have a cls._get_by_name(ado_client, payload["name"])
-    #     # We do, however, have to be able to compare a ado_wrapper resource to a payload, that's much tricker...
-    #     # Perhaps we can use the editable resources, since if we can't edit it, what's the point in "maintaining" it?
-    #     # Also have to be able to update a resource based on the ado_wrapper attributes.
-    #     # Another problem is things running in this class, need to be able to call Resource._create, not StateManagedResource.create
-    #     # Since this might call create, we don't have the payload...
-
-    #     # To be able to get the attributes, we need to format an object, but this isn't going to work.
-    #     # Instead, I suppose we could add a new function, which takes the input args and creates an object#
-    #     # From that data, basically a mapping of input params -> object (fake), then we can do getattr(x) get_internal_field_names
-
-    #     existing_case = cls.get_by_name(ado_client, args[0])  # type: ignore[abc]  All maintainable resources have this.
-    #     print(f"1. {existing_case=}")
-    #     if existing_case:
-    #         print("2. Existing case is Truthy, detecting changes")
-    #         editable_attributes = get_editable_fields(existing_case)
-    #         existing_dictionary = {x: getattr(existing_case, x) for x in editable_attributes}
-    #         potenatial_resource = cls.from_create_args(*(args+kwargs.values()))  # type: ignore[abc]
-    #         potential_dictionary = {x: getattr(potenatial_resource, x) for x in editable_attributes}
-    #         differences = {key: potential_dictionary[key] for key in potential_dictionary
-    #                       if potential_dictionary[key] != existing_dictionary[key]}
-    #         if differences:
-    #             print("3. Changes detected, updating resource")
-    #             for key, value in differences.items():
-    #                 existing_case.update(ado_client, key, value)
-    #                 setattr(existing_case, key, value)
-    #             good_resource = existing_case
-    #         else:
-    #             print("3. No differences detected, doing nothing.")
-    #             good_resource = existing_case
-    #     else:
-    #         print("2. Existing case is Falsey, creating")
-    #         good_resource = cls.create(ado_client, url, payload)  # type: ignore
-    #     print("4. Returning")
-    #     return good_resource
-
     @classmethod
     def _delete_by_id(cls: Type[T], ado_client: "AdoClient", url: str, resource_id: str) -> None:
         """Deletes an object by its id. The id is passed so it can be removed from state"""
@@ -211,6 +174,21 @@ class StateManagedResource:
         return [cls.from_request_payload(resource) for resource in request.json()["value"]]
 
     @classmethod
+    def _get_all_paginated(cls: Type[T], ado_client: "AdoClient", url: str, limit: int | None = None, page_size: int = 1000) -> list[T]:
+        """Gets all resources for a url, paginated"""
+        fetched_resourced: list[T] = []
+        skip = 0
+        while True:
+            resources = cls._get_all(ado_client, url + f"&$skip={skip}")
+            if not resources:
+                return fetched_resourced
+            for resource in resources:
+                if limit is not None and len(fetched_resourced) >= limit:
+                    return fetched_resourced
+                fetched_resourced.append(resource)
+            skip += page_size
+
+    @classmethod
     def _get_by_abstract_filter(cls: Type[T], ado_client: "AdoClient", func: Callable[[T], bool]) -> T | None:
         """Used internally for getting resources by a filter function. The function should return True if the resource is the one you want."""
         resources = cls.get_all(ado_client)  # type: ignore[attr-defined]  # pylint: disable=no-value-for-parameter, no-member
@@ -218,6 +196,47 @@ class StateManagedResource:
             if func(resource):
                 return resource  # type: ignore[no-any-return]
         return None
+
+    # @classmethod
+    # def maintain(cls: Type[T], ado_client: "AdoClient", *args: Any, **kwargs: Any) -> T:  # url: str, payload: dict[str, Any]
+    #     return  # type: ignore
+    #     # To be able to get the existing one, we need to be able to fetch it by name
+    #     # We can assume that all the resources have a cls._get_by_name(ado_client, payload["name"])
+    #     # We do, however, have to be able to compare a ado_wrapper resource to a payload, that's much tricker...
+    #     # Perhaps we can use the editable resources, since if we can't edit it, what's the point in "maintaining" it?
+    #     # Also have to be able to update a resource based on the ado_wrapper attributes.
+    #     # Another problem is things running in this class, need to be able to call Resource._create, not StateManagedResource.create
+    #     # Since this might call create, we don't have the payload...
+
+    #     # To be able to get the attributes, we need to format an object, but this isn't going to work.
+    #     # Instead, I suppose we could add a new function, which takes the input args and creates an object#
+    #     # From that data, basically a mapping of input params -> object (fake), then we can do getattr(x) get_internal_field_names
+
+    #     # All maintainable resources have this.
+    #     existing_case = cls.get_by_name(ado_client, args[0])
+    #     print(f"1. {existing_case=}")
+    #     if existing_case:
+    #         print("2. Existing case is Truthy, detecting changes")
+    #         editable_attributes = get_editable_fields(existing_case)
+    #         existing_dictionary = {x: getattr(existing_case, x) for x in editable_attributes}
+    #         potenatial_resource = cls.from_create_args(*(args + kwargs.values()))
+    #         potential_dictionary = {x: getattr(potenatial_resource, x) for x in editable_attributes}
+    #         differences = {key: potential_dictionary[key] for key in potential_dictionary
+    #                        if potential_dictionary[key] != existing_dictionary[key]}
+    #         if differences:
+    #             print("3. Changes detected, updating resource")
+    #             for key, value in differences.items():
+    #                 existing_case.update(ado_client, key, value)
+    #                 setattr(existing_case, key, value)
+    #             good_resource = existing_case
+    #         else:
+    #             print("3. No differences detected, doing nothing.")
+    #             good_resource = existing_case
+    #     else:
+    #         print("2. Existing case is Falsey, creating")
+    #         good_resource = cls.create(ado_client)
+    #     print("4. Returning")
+    #     return good_resource
 
     # def set_lifecycle_policy(self, ado_client: "AdoClient", policy: Literal["prevent_destroy", "ignore_changes"]) -> None:
     #     self.life_cycle_policy = policy  # TODO